add: readContractData
