+++ package.json
@@ -11,7 +11,7 @@
   "main": "index.js",
   "scripts": {
     "start:build": "tsc -w",
-    "start:run": "nodemon dist/index.js",
+    "start:ts-template": "nodemon dist/ts-template/index.js",
     "start": "concurrently npm:start:*",
     "start:trade-service": "tsc -b && nodemon dist/trade-service/index.js"
   },
