+++ src/api-template/auth/emailAuthenticate_template.ts
