+++ src/api-template/utils/generateClientSignature.ts
