+++ src/api-template/utils/generateSignMessagePayload.ts
