+++ src/api-template/utils/generateUserOp.ts
@@ -1,42 +1,55 @@
 import {
-    encodeAbiParameters,
-    keccak256,
-    pad,
-    parseAbiParameters,
-    hexToBigInt,
-    type Hex,
-    concat,
+  encodeAbiParameters,
+  keccak256,
+  pad,
+  parseAbiParameters,
+  hexToBigInt,
+  type Hex,
+  concat,
 } from "viem";
 import { Constants } from "../helper/constants.js";
 import type { UserOp, PackedUserOp, Hash } from "../helper/types.js";
 /**
  * Creates the Packed UserOp (User Operation)
- * 
+ *
  * This function packages various user operation parameters into a structured format.
- * 
- * @param userOp - Object containing the user operation details. 
+ *
+ * @param userOp - Object containing the user operation details.
  * @returns Formatted UserOp object with packed gas parameters
  * @throws Error if any required parameters are missing
  */
 export function generatePackedUserOp(userOp: UserOp): PackedUserOp {
-    if (!userOp.sender || !userOp.nonce || !userOp.callData || !userOp.preVerificationGas || !userOp.verificationGasLimit || !userOp.callGasLimit || !userOp.maxFeePerGas || !userOp.maxPriorityFeePerGas || userOp.paymaster == void 0 || !userOp.paymasterVerificationGasLimit || !userOp.paymasterPostOpGasLimit || userOp.paymasterData == void 0) {
-        throw new Error("Invalid UserOp");
-    }
-     const accountGasLimits: Hash = ('0x' +
+  if (
+    !userOp.sender ||
+    !userOp.nonce ||
+    !userOp.callData ||
+    !userOp.preVerificationGas ||
+    !userOp.verificationGasLimit ||
+    !userOp.callGasLimit ||
+    !userOp.maxFeePerGas ||
+    !userOp.maxPriorityFeePerGas ||
+    userOp.paymaster == void 0 ||
+    !userOp.paymasterVerificationGasLimit ||
+    !userOp.paymasterPostOpGasLimit ||
+    userOp.paymasterData == void 0
+  ) {
+    throw new Error("Invalid UserOp");
+  }
+  const accountGasLimits: Hash = ("0x" +
     pad(userOp.verificationGasLimit, {
       size: 16,
-    }).replace('0x', '') +
+    }).replace("0x", "") +
     pad(userOp.callGasLimit, {
       size: 16,
-    }).replace('0x', '')) as Hash;
+    }).replace("0x", "")) as Hash;
 
-  const gasFees: Hash = ('0x' +
+  const gasFees: Hash = ("0x" +
     pad(userOp.maxFeePerGas, {
       size: 16,
-    }).replace('0x', '') +
+    }).replace("0x", "") +
     pad(userOp.maxPriorityFeePerGas, {
       size: 16,
-    }).replace('0x', '')) as Hash;
+    }).replace("0x", "")) as Hash;
 
   const paymasterAndData = userOp.paymaster
     ? concat([
@@ -49,12 +62,12 @@ export function generatePackedUserOp(userOp: UserOp): PackedUserOp {
         }),
         userOp.paymasterData,
       ])
-    : '0x';
+    : "0x";
 
   const packedUserOp: PackedUserOp = {
     sender: userOp.sender,
     nonce: userOp.nonce,
-    initCode: '0x',
+    initCode: "0x",
     callData: userOp.callData,
     preVerificationGas: userOp.preVerificationGas,
     accountGasLimits,
@@ -66,17 +79,17 @@ export function generatePackedUserOp(userOp: UserOp): PackedUserOp {
 }
 
 /**
- * Generates the userOp Hash 
+ * Generates the userOp Hash
  * Creates a unique hash that identifies the user operation.
  * This hash is used for signing purpose.
- * 
+ *
  * @param userOp - Packed user operation object (output from generatePackedUserOp)
  * @returns The keccak256 hash of the user operation.
  */
-export function generateUserOpHash(userOp: PackedUserOp ): Hash {
-   const pack = encodeAbiParameters(
+export function generateUserOpHash(userOp: PackedUserOp): Hash {
+  const pack = encodeAbiParameters(
     parseAbiParameters(
-      'address, bytes32, bytes32, bytes32, bytes32, uint256, bytes32, bytes32',
+      "address, bytes32, bytes32, bytes32, bytes32, uint256, bytes32, bytes32"
     ),
     [
       userOp.sender,
@@ -99,18 +112,17 @@ export function generateUserOpHash(userOp: PackedUserOp ): Hash {
       pad(keccak256(userOp.paymasterAndData), {
         size: 32,
       }),
-    ],
+    ]
   );
 
-
-   const userOpPack = encodeAbiParameters(
-        parseAbiParameters("bytes32, address, uint256"),
-        [
-            keccak256(pack),
-            Constants.ENV_CONFIG.SANDBOX.ENTRYPOINT_CONTRACT_ADDRESS,
-            BigInt(Constants.ENV_CONFIG.SANDBOX.CHAIN_ID)
-        ]
-    );
+  const userOpPack = encodeAbiParameters(
+    parseAbiParameters("bytes32, address, uint256"),
+    [
+      keccak256(pack),
+      Constants.ENV_CONFIG.SANDBOX.ENTRYPOINT_CONTRACT_ADDRESS,
+      BigInt(Constants.ENV_CONFIG.SANDBOX.CHAIN_ID),
+    ]
+  );
 
   return keccak256(userOpPack);
 }
