+++ src/api-template/utils/getAuthorizationToken.ts
